Refactor entire code base to capture linter output using JSON
Older versions of golangci-lint were sending output exclusively to
/dev/stdout and using a more standard format. However, the latest version
either has a bug or is changing from its original design, because there
are errors and warnings in /dev/stderr too which differ from the other
messages just by a few characters, for example, missing a colon after the
column number.

I’ve fixed these problems and improved the parser by leveraging the JSON
output, and formalizing both /dev/stderr as Errors and /dev/stdout as
warnings.

I also went ahead and submitted the project to SublimeLinter repository
for inclusion in their registry to allow people to install the linter
via Package Control [1]. I immediately approached the original author
to ask for a review and hopefully merge the fixes so others can benefit
from the improvements.

[1] https://github.com/SublimeLinter/package_control_channel/pull/83